pyproject.toml
[project]
name = "financial-auditor"
version = "0.1.0"
description = "Local-first AI financial document auditor with auditable extraction, validation, confidence routing, and reporting."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.9",
    "PyYAML>=6.0.1",
    "groq>=0.9.0",
    "unstructured[pdf]>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["financial_auditor*"]
